requirements.txt
flask-3.0.2
flask_sqlalchemy-3.1.1
Flask_Migrate-4.0.7
pillow-10.2.0
numpy-1.26.4
opencv_python-4.9.0
joblib-1.3.2
scipy-1.12.0
scikit-learn-1.4.1
matplotlib-3.8.3
pandas-2.2.1
tqdm